move raw bullet code from interface to physics lib
